// lib/api/generated/models/wikiArticleListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { DepartmentRef } from './departmentRef';
import type { UserShortResponse } from './userShortResponse';

export interface WikiArticleListResponse {
  author?: UserShortResponse;
  categoryName?: string;
  departments?: DepartmentRef[];
  excerpt?: string;
  id?: number;
  likeCount?: number;
  likedByCurrentUser?: boolean;
  slug?: string;
  tags?: string[];
  title?: string;
  updatedAt?: string;
  viewCount?: number;
}